Add x64 support
