+++ extension/AMBuilder
@@ -6,9 +6,15 @@ projectname = 'connect'
 project = builder.LibraryProject(projectname + '.ext')
 project.sources = [
   'extension.cpp',
-  'asm/asm.c',
-  'CDetour/detours.cpp',
-  'sdk/smsdk_ext.cpp'
+  os.path.join(Extension.sm_root, 'public', 'smsdk_ext.cpp'),
+  os.path.join(Extension.sm_root, 'public', 'CDetour', 'detours.cpp'),
+  os.path.join(Extension.sm_root, 'public', 'asm', 'asm.c'),
+  os.path.join(Extension.sm_root, 'public', 'libudis86', 'decode.c'),
+  os.path.join(Extension.sm_root, 'public', 'libudis86', 'itab.c'),
+  os.path.join(Extension.sm_root, 'public', 'libudis86', 'syn-att.c'),
+  os.path.join(Extension.sm_root, 'public', 'libudis86', 'syn-intel.c'),
+  os.path.join(Extension.sm_root, 'public', 'libudis86', 'syn.c'),
+  os.path.join(Extension.sm_root, 'public', 'libudis86', 'udis86.c')
 ]
 
 for sdk_name in Extension.sdks:
